move usePublicTaxonomies to shared utils file
